+++ app/src/main/java/com/boardgamegeek/io/UserRequest.java
@@ -11,28 +11,16 @@ import timber.log.Timber;
 public class UserRequest {
 	private final BggService bggService;
 	private final String name;
-	private final boolean includeBuddies;
 
 	public UserRequest(BggService service, String name) {
 		this.bggService = service;
 		this.name = name;
-		this.includeBuddies = false;
-	}
-
-	public UserRequest(BggService service, String name, boolean includeBuddies) {
-		this.bggService = service;
-		this.name = name;
-		this.includeBuddies = includeBuddies;
 	}
 
 	public User execute() {
 		Call<User> call;
 		try {
-			if (includeBuddies) {
-				call = bggService.user(name, 1, 1);
-			} else {
-				call = bggService.user(name);
-			}
+			call = bggService.user(name);
 			Response<User> response = call.execute();
 			if (response.isSuccessful()) {
 				return response.body();