+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Release Notes
 =============
 
+Version 6.3.5
+-------------
+ * Clear database and re-sync play information
+ * Syncing improvements
+ * Bug fixes
+
 Version 6.3.4
 -------------
  * Bug fixes